refactor: rename package to codidiary (lowercase)
