+++ app/src/main/java/ddwu/com/mobile/wearly_frontend/codidiary/data/CodiCategoryResponse.kt
@@ -1,4 +1,4 @@
-package ddwu.com.mobile.wearly_frontend.codiDiary.data
+package ddwu.com.mobile.wearly_frontend.codidiary.data
 
 data class CodiCategoryResponse(
     val success: Boolean,
