+++ app/src/main/java/ddwu/com/mobile/wearly_frontend/codidiary/ui/CodiDiaryWriteFragment.kt
@@ -1,5 +1,7 @@
-package ddwu.com.mobile.wearly_frontend.codiDiary.ui
+package ddwu.com.mobile.wearly_frontend.codidiary.ui
 
+import CodiDiaryRecordRequest
+import ddwu.com.mobile.wearly_frontend.R
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
@@ -9,27 +11,25 @@ import android.widget.FrameLayout
 import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.engine.DiskCacheStrategy
 import com.bumptech.glide.request.RequestOptions
 import com.google.android.material.snackbar.Snackbar
-import ddwu.com.mobile.wearly_frontend.R
 import ddwu.com.mobile.wearly_frontend.TokenManager
-import ddwu.com.mobile.wearly_frontend.codiDiary.data.CodiDiaryEditRequest
-import ddwu.com.mobile.wearly_frontend.codiDiary.data.viewmodel.CodiDiaryViewModel
+import ddwu.com.mobile.wearly_frontend.codidiary.data.viewmodel.CodiDiaryViewModel
 import ddwu.com.mobile.wearly_frontend.databinding.FragmentCodiDiaryBinding
 
-class CodiDiaryEditFragment : Fragment() {
+class CodiDiaryWriteFragment: Fragment() {
 
     private lateinit var binding: FragmentCodiDiaryBinding
-    private val codiDiaryViewModel: CodiDiaryViewModel by activityViewModels()
+
+    private val codiDiaryWriteViewModel: CodiDiaryViewModel by viewModels()
 
     private var isLiked = false
-    private var dateId: Int = -1
 
+    // 카테고리별 배치 좌표 맵 (중복 제거 및 가독성을 위해 상단 배치)
     private val categoryLayoutMap = mapOf(
         "아우터" to Triple(0.25f, 0.25f, 1),
         "상의" to Triple(0.65f, 0.35f, 2),
@@ -54,39 +54,44 @@ class CodiDiaryEditFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        dateId = arguments?.getInt("dateId") ?: -1
-        val outfitName = arguments?.getString("outfitName")
-        val memo = arguments?.getString("memo")
-        isLiked = arguments?.getBoolean("isHeart") ?: false
-        val wearDate = arguments?.getString("wearDate")
-        val iconCode = arguments?.getString("weatherIcon")?.toIntOrNull() ?: 0
-        val tempMin = arguments?.getDouble("tempMin") ?: 0.0
-        val tempMax = arguments?.getDouble("tempMax") ?: 0.0
 
-        val categories = arguments?.getStringArray("selectedClothCategories") ?: arrayOf()
-        val images = arguments?.getStringArray("selectedClothImages") ?: arrayOf()
+        val selectedDate = arguments?.getString("selectedDate")
+        binding.diaryDayTv.setText(selectedDate)
 
+        val serverDate = selectedDate?.replace("년 ", "-")
+            ?.replace("월 ", "-")
+            ?.replace("일", "")
+            ?.split("-")
+            ?.let { parts ->
+                val y = parts[0]
+                val m = parts[1].padStart(2, '0')
+                val d = parts[2].trim().padStart(2, '0')
+                "$y-$m-$d"
+            } ?: ""
 
-
-        if (wearDate != null) {
-            val dateParts = wearDate.split("-")
-            if (dateParts.size == 3) {
-                val formattedDate = "${dateParts[0]}년 ${dateParts[1].toInt()}월 ${dateParts[2].toInt()}일"
-                binding.diaryDayTv.text = formattedDate
-            } else {
-                binding.diaryDayTv.text = wearDate
-            }
-        }
-        binding.diaryTitleEt.setText(outfitName)
-        binding.diaryEt.setText(memo)
-        binding.diaryTempTv.text = "${tempMin.toInt()}° / ${tempMax.toInt()}°"
+        val iconCode = arguments?.getInt("weatherIcon") ?: 0
         binding.diaryWeatherIcon.setImageResource(getWeatherDrawable(iconCode))
 
-        updateLikeUI()
+        val temperature = arguments?.getString("temperature") ?: ""
+        binding.diaryTempTv.setText(temperature)
 
-        if (categories.isNotEmpty() && images.isNotEmpty()) {
-            displayOutfits(categories, images)
+        val temps = temperature.replace("°", "").split("/")
+        val minTemp = temps.getOrNull(0)?.toIntOrNull() ?: 0
+        val maxTemp = temps.getOrNull(1)?.toIntOrNull() ?: 0
+
+        val selectedIds = arguments?.getIntArray("selectedClothIds") ?: intArrayOf(0)
+
+        // 이전 프래그먼트에서 넘겨준 카테고리 및 이미지 정보 받기
+        val selectedCategories = arguments?.getStringArray("selectedClothCategories")
+        val selectedImages = arguments?.getStringArray("selectedClothImages")
+
+        Log.d("DiaryWrite", "선택된 옷 ID들: ${selectedIds?.contentToString()}")
+
+        // 이미지 배치 함수 호출
+        if (selectedCategories != null && selectedImages != null) {
+            displaySelectedOutfits(selectedCategories, selectedImages)
         }
+
 
         // --------------- 리스너 ---------------
 
@@ -95,70 +100,72 @@ class CodiDiaryEditFragment : Fragment() {
             findNavController().popBackStack()
         }
 
-        // 좋아요 토글
-        binding.diaryLikeBtnToggle.setOnClickListener {
-            isLiked = !isLiked
-            updateLikeUI()
-        }
-
-        // 수정 완료 (PATCH API 호출)
+        // 다이어리 저장
         binding.diarySubmitBtn.setOnClickListener {
             val title = binding.diaryTitleEt.text.toString()
-            val content = binding.diaryEt.text.toString()
+            val diary = binding.diaryEt.text.toString()
 
             if (title.isEmpty()) {
-                Snackbar.make(binding.root, "제목을 입력해주세요!", Snackbar.LENGTH_SHORT).show()
+                binding.diaryTitleEt.hint = "제목을 입력해주세요!!"
+                binding.diaryEt.hint = "일기를 입력해주세요!!"
+
                 return@setOnClickListener
             }
+            if (selectedDate.isNullOrEmpty()) {
+                findNavController().popBackStack()
+                Snackbar.make(binding.root, "오류가 발생했습니다. 다시 시도해주세요.", Snackbar.LENGTH_SHORT).show()
+            }
 
-            val updateRequest = CodiDiaryEditRequest(
+
+            val request = CodiDiaryRecordRequest(
+                wear_date = serverDate,
+                clothes_ids = selectedIds.toList(),
                 outfit_name = title,
-                memo = content,
-                is_heart = isLiked,
-                wear_date = wearDate,
-                temp_min = tempMin,
-                temp_max = tempMax,
-                weather_icon = iconCode.toString()
+                temp_min = minTemp,
+                temp_max = maxTemp,
+                weather_icon = iconCode.toString(),
+                memo = diary,
+                is_heart = isLiked
             )
 
             val token = TokenManager(requireContext()).getToken()
-            if (token != null && dateId != -1) {
-                codiDiaryViewModel.updateRecord(token, dateId, updateRequest)
-            }
-        }
 
-        codiDiaryViewModel.updateStatus.observe(viewLifecycleOwner) { isSuccess ->
-            if (isSuccess) {
-                val token = TokenManager(requireContext()).getToken()
-                val wearDate = arguments?.getString("wearDate")
-
-                if (token != null && wearDate != null) {
-                    codiDiaryViewModel.fetchDiaryRead(token, wearDate)
-
-                    Snackbar.make(binding.root, "기록이 수정되었습니다.", Snackbar.LENGTH_SHORT).show()
-                    findNavController().popBackStack()
-                }
+            if (!token.isNullOrEmpty()){
+                codiDiaryWriteViewModel.saveRecord(token, isWeatherLog = true, request = request)
             } else {
-                Snackbar.make(binding.root, "수정에 실패했습니다.", Snackbar.LENGTH_SHORT).show()
+            }
+        }
+
+        codiDiaryWriteViewModel.saveStatus.observe(viewLifecycleOwner) { isSuccess ->
+            if (isSuccess) {
+                Snackbar.make(binding.root, "코디 일기가 저장되었습니다.", Snackbar.LENGTH_SHORT).show()
+                findNavController().popBackStack(R.id.calendarFragment, false)
+
+            } else {
+                Snackbar.make(binding.root, "저장에 실패했습니다. 다시 시도해주세요.", Snackbar.LENGTH_SHORT).show()
+            }
+        }
+
+
+        // 좋아요
+        binding.diaryLikeBtnToggle.setOnClickListener {
+            isLiked = !isLiked
+
+            if (isLiked){
+                binding.diaryLikeSelected.visibility = View.VISIBLE
+                binding.diaryLikeUnselected.visibility = View.GONE
+            }
+            else {
+                binding.diaryLikeUnselected.visibility = View.VISIBLE
+                binding.diaryLikeSelected.visibility = View.GONE
             }
         }
     }
-
-    private fun updateLikeUI() {
-        if (isLiked) {
-            binding.diaryLikeSelected.visibility = View.VISIBLE
-            binding.diaryLikeUnselected.visibility = View.GONE
-        } else {
-            binding.diaryLikeUnselected.visibility = View.VISIBLE
-            binding.diaryLikeSelected.visibility = View.GONE
-        }
-    }
-
 
     /**
-     * 이미지 배치 함수
+     * 선택된 이미지들을 FrameLayout에 배치하는 함수
      */
-    private fun displayOutfits(categories: Array<String>, images: Array<String>) {
+    private fun displaySelectedOutfits(categories: Array<String>, images: Array<String>) {
         val container = binding.diaryClothesFrame
         container.removeAllViews()
 
@@ -198,6 +205,9 @@ class CodiDiaryEditFragment : Fragment() {
         }
     }
 
+    /**
+     * 날씨 코드를 아이콘으로 변환
+     */
     private fun getWeatherDrawable(iconCode: Int): Int {
         return when (iconCode) {
             0 -> R.drawable.ic_weather_sunny
@@ -208,13 +218,17 @@ class CodiDiaryEditFragment : Fragment() {
         }
     }
 
+
     override fun onResume() {
         super.onResume()
+
+        // 액션바 숨기기
         (activity as AppCompatActivity).supportActionBar?.hide()
     }
 
     override fun onPause() {
         super.onPause()
+        // 다른 화면으로 나갈 때 다시 보이게 하기
         (activity as AppCompatActivity).supportActionBar?.show()
     }
 }
